requirements.txt
accelerate==1.12.0
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
audioread==3.1.0
autodistill==0.1.29
autodistill-grounding-dino==0.1.4
av==16.1.0
babel==2.17.0
beautifulsoup4==4.14.2
bithuman==0.6.7
bitsandbytes==0.49.1
black==26.1.0
bleach==6.3.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
cut-cross-entropy==25.1.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.3.0
debugpy==1.8.17
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.36.0
dill==0.4.0
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
google==3.0.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-speech==2.35.0
google-cloud-texttospeech==2.33.0
google-genai==1.57.0
googleapis-common-protos==1.72.0
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.16
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
librosa==0.11.0
livekit==1.0.23
livekit-agents==1.3.10
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-assemblyai==1.3.10
livekit-plugins-bey==1.3.10
livekit-plugins-bithuman==1.3.10
livekit-plugins-cartesia==1.3.10
livekit-plugins-deepgram==1.3.10
livekit-plugins-elevenlabs==1.3.10
livekit-plugins-google==1.3.10
livekit-plugins-groq==1.3.10
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.3.10
livekit-plugins-rime==1.3.10
livekit-plugins-silero==1.3.10
livekit-plugins-speechmatics==1.3.10
livekit-plugins-tavus==1.3.10
livekit-plugins-turn-detector==1.3.10
livekit-protocol==1.1.1
llvmlite==0.46.0
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.25.0
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
notebook==7.5.0
notebook_shim==0.2.4
numba==0.63.1
numpy==2.2.6
nvidia-cublas==13.0.0.19
nvidia-cuda-cupti==13.0.48
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-runtime==13.0.48
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.3.29
nvidia-cusparse==12.6.2.49
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
onnxruntime==1.23.1
openai==2.15.0
openai-agents==0.6.5
opencv-python==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==1.0.3
peft==0.18.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
pre_commit==4.5.1
premailer==3.10.0
proglog==0.1.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycocotools==2.0.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytokens==0.4.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==26.4.0
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rf-segment-anything==1.0
rf_groundingdino==0.1.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
ruff==0.14.13
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
soxr==0.5.0.post1
speechmatics-rt==0.5.3
srt==3.5.3
sse-starlette==3.1.2
stack-data==0.6.3
starlette==0.50.0
supervision==0.27.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.24
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.1+cu130
torchao==0.15.0
torchvision==0.24.1+cu130
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.1
trl==0.24.0
typeguard==4.4.4
typer==0.21.1
typer-slim==0.20.0
types-protobuf==6.32.1.20251210
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.5
tzdata==2025.2
unsloth==2026.1.4
unsloth_zoo==2026.1.4
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.40.0
virtualenv==20.36.1
vosk==0.3.45
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.46.3
xformers==0.0.33.post2
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
